+++ SpaceSim/Graphics/EffectCache.h
@@ -22,10 +22,10 @@ public:
     const Effect* EffectCache::createEffect(Resource* resource, const std::string& resourceFileName);
 
     const Effect* getEffect(const std::string& name) const;
-    const Effect* getEffect(unsigned int effectHash) const;
+    const Effect* getEffect(size_t effectHash) const;
     void addEffect(const std::string& name, const Effect& effect);
 private:
 
-    std::map<unsigned int, Effect> m_effects;
+    std::map<size_t, Effect> m_effects;
 };
