+++ SpaceSim/Input/InputAction.h
@@ -37,15 +37,15 @@ namespace InputActions
         ActionType(const std::string& actionTypeName) : m_actionType(hashString(actionTypeName)), m_actionTypeLngKey(0) {}
         ActionType(const std::string& actionTypeName, const std::string& lngActionName) : m_actionType(hashString(actionTypeName)), m_actionTypeLngKey(hashString(lngActionName.c_str())) {}
 
-        unsigned int getType() const { return m_actionType; }
-        unsigned int getLanguageString() const { return m_actionTypeLngKey; }
+        size_t getType() const { return m_actionType; }
+        size_t getLanguageString() const { return m_actionTypeLngKey; }
 #if defined( DEBUG ) || defined( _DEBUG )
         //const std::string& getNameString() const { return m_actionTypeName; }
 #endif
         bool operator < ( const ActionType& rhs ) const { return m_actionType < rhs.m_actionType; }
     private:
-        unsigned int m_actionType;
-        unsigned int m_actionTypeLngKey;
+        size_t m_actionType;
+        size_t m_actionTypeLngKey;
         //std::string m_actionTypeName;
     };
 }
